fix(idenfy): isolate webhook signature verifier from init.ts
The webhooks.test.ts imported verifyIdenfyWebhookSignature from
webhooks.ts, which transitively imports init.ts and runs validateEnv()
at module load — failing CI where HOLONYM_ISSUER_PRIVKEY is unset.

Extract the pure HMAC verifier to webhook-signature.ts (no init.ts
dependency); webhooks.ts re-exports it for existing callers.
